// src/app/models/Details.ts
export class Details {

    currentPage: number | undefined;
    hasMore: boolean | undefined;
    next: string | undefined;
    previous: string | undefined;
    pageCount: number | undefined;
    pageSelector: string | undefined;
    perPage: number | undefined;
    segment: number | undefined;
    total: number | undefined;
}